+++ src/vaadin-combo-box.d.ts
@@ -239,4 +239,3 @@ declare global {
 }
 
 export { ComboBoxElement };
-
