+++ vaadin-combo-box.d.ts
@@ -1,2 +1,2 @@
 export * from './src/vaadin-combo-box.js';
-export * from './@types/interfaces';
+export * from './src/interfaces';
